Implemented TaskManager class by refactoring methods
