--- src/main/java/seedu/duke/TaskManager.java
@@ -1,4 +1,0 @@
-package seedu.duke;
-
-public class TaskManager {
-}
