+++ src/main/java/seedu/duke/task/FinalExam.java
@@ -1,9 +1,9 @@
-package seedu.duke;
+package seedu.duke.task;
 
-public class Midterm extends Task {
+public class FinalExam extends Task {
     protected String on;
 
-    public Midterm(String module, String description, String message, String on) {
+    public FinalExam(String module, String description, String message, String on) {
         super(module, description, message);
         this.on = on;
     }
