+++ src/test/java/seedu/duke/TaskListTest.java
@@ -1,6 +1,7 @@
 package seedu.duke;
 
 import org.junit.jupiter.api.Test;
+import seedu.duke.task.TaskList;
 
 import java.time.format.DateTimeParseException;
